evapi: Removed cast of bind port to short, as this restricted ports to signed 16-bit int values (< 32768).
htons() takes unsigned 16-bit integer and returns the same. When this value was cast to (short), higher ports could not be used.

(cherry picked from commit e9545ae1e490ff768270d03ea28f14b751d2c132)
(cherry picked from commit 4935d3c3c9ff76f34670fe39e26df11f909d6453)

+++ src/modules/evapi/evapi_dispatch.c
@@ -676,7 +676,7 @@ int evapi_run_dispatcher(char *laddr, int lport)
 
 	bzero(&evapi_srv_addr, sizeof(evapi_srv_addr));
 	evapi_srv_addr.sin_family = h->h_addrtype;
-	evapi_srv_addr.sin_port   = htons((short)lport);
+	evapi_srv_addr.sin_port   = htons(lport);
 	evapi_srv_addr.sin_addr  = *(struct in_addr*)h->h_addr;
 
 	/* Set SO_REUSEADDR option on listening socket so that we don't
